refactor(orchestrator): consolidate report types into report.rs
ConfigureReport, MonitoringReport and LogsReport were defined in their
respective phase files (prepare.rs, monitoring.rs, logs.rs). Move them
into a dedicated report.rs so all public output types are discoverable
in one place.

Add TickReport (future return type of Orchestrator::tick(), issue #172)
alongside the existing reports. MetricsTick carries Option<String> (the
JSON-serialised BenchmarkResults) so the caller holds the data without
pulling in the <N, C> type parameters; FaultUpdate carries a
CrashRecoveryAction.

Add BenchmarkResults::to_json() in collector.rs for use by tick() when
it serialises results into TickReport::MetricsTick.

lib.rs gains pub mod report. orchestrator.rs re-exports via
pub use crate::report::{ConfigureReport, MonitoringReport} so existing
import paths are unchanged.
